Fix FileSize being int which is too small for plot sizes

+++ Chia.NET/Models/Harvester/Plot.cs
@@ -5,7 +5,7 @@ namespace Chia.NET.Models
     public sealed class Plot
     {
         [JsonPropertyName("file_size")]
-        public int FileSize { get; init; }
+        public ulong FileSize { get; init; }
 
         [JsonPropertyName("filename")]
         public string FileName { get; init; }
